+++ src/components/scenes/Staking/StakeModifyScene.tsx
@@ -7,10 +7,11 @@ import { sprintf } from 'sprintf-js'
 
 import { lstrings } from '../../../locales/strings'
 import { ChangeQuote, ChangeQuoteRequest, QuoteAllocation, StakeBelowLimitError, StakePoolFullError } from '../../../plugins/stake-plugins/types'
-import { getDenominationFromCurrencyInfo, getDisplayDenomination } from '../../../selectors/DenominationSelectors'
+import { getDisplayDenomination, getExchangeDenominationFromAccount } from '../../../selectors/DenominationSelectors'
 import { useSelector } from '../../../types/reactRedux'
 import { EdgeSceneProps } from '../../../types/routerTypes'
 import { getCurrencyIconUris } from '../../../util/CdnUris'
+import { getTokenId } from '../../../util/CurrencyInfoHelpers'
 import { getWalletName } from '../../../util/CurrencyWalletHelpers'
 import { getPolicyIconUris, getPolicyTitleName, getPositionAllocations, getUnstakeText } from '../../../util/stakeUtils'
 import { toBigNumberString } from '../../../util/toBigNumberString'
@@ -279,7 +280,7 @@ const StakeModifySceneComponent = (props: Props) => {
         : undefined
 
     const quoteCurrencyCode = currencyCode
-    const quoteDenom = getDenominationFromCurrencyInfo(wallet.currencyInfo, quoteCurrencyCode)
+    const quoteDenom = getExchangeDenominationFromAccount(account, pluginId, quoteCurrencyCode)
 
     const title =
       allocationType === 'stake'
@@ -319,14 +320,16 @@ const StakeModifySceneComponent = (props: Props) => {
         : undefined
     if (quoteAllocation == null) return null
 
-    const quoteDenom = getDenominationFromCurrencyInfo(wallet.currencyInfo, currencyCode)
+    const quoteDenom = getExchangeDenominationFromAccount(account, pluginId, currencyCode)
     const title = modification === 'stake' ? lstrings.stake_estimated_staking_fee : lstrings.stake_estimated_unstaking_fee
+    const tokenId = getTokenId(account, pluginId, currencyCode)
 
     return (
       <CryptoFiatAmountTile
         type="questionable"
         title={title}
         nativeCryptoAmount={quoteAllocation?.nativeAmount ?? '0'}
+        tokenId={tokenId}
         walletId={wallet.id}
         denomination={quoteDenom}
         onPress={handlePressStakingFee(modification)}
@@ -345,13 +348,15 @@ const StakeModifySceneComponent = (props: Props) => {
         : undefined
     if (quoteAllocation == null) return null
 
-    const quoteDenom = getDenominationFromCurrencyInfo(wallet.currencyInfo, currencyCode)
+    const quoteDenom = getExchangeDenominationFromAccount(account, pluginId, currencyCode)
+    const tokenId = getTokenId(account, pluginId, currencyCode)
 
     return (
       <CryptoFiatAmountTile
         type="questionable"
         title={lstrings.stake_future_unstaking_fee}
         nativeCryptoAmount={quoteAllocation?.nativeAmount ?? '0'}
+        tokenId={tokenId}
         walletId={wallet.id}
         denomination={quoteDenom}
         onPress={handlePressFutureUnstakingFee}